Trekkjer ut spesial-strategiane våre også ut i eigne filer. Dette er kanskje ei smakssak, men eg synes det er fint å ha kvar klasse i si eiga fil
