+++ src/main/kotlin/no/nav/familie/ba/sak/config/featureToggle/ByClusterStrategy.kt
@@ -1,0 +1,13 @@
+package no.nav.familie.ba.sak.config.featureToggle
+
+import io.getunleash.strategy.Strategy
+
+class ByClusterStrategy(private val clusterName: String) : Strategy {
+
+    override fun isEnabled(parameters: MutableMap<String, String>): Boolean {
+        if (parameters.isEmpty()) return false
+        return parameters["cluster"]?.contains(clusterName) ?: false
+    }
+
+    override fun getName(): String = "byCluster"
+}
